Add fail safe in case app crashes.
When the app crashes, ActiveJob needs to what time the job was supposed
to have been performed. The message_expires_at variable takes care of
that. Also, it has to be serialized and then deserialized because
ActiveJob cannot accept Time parameters.

Method has been changed from timer to create to be RESTful.